AzurePipeline for github/requirements.txt
azure-ai-textanalytics==5.2.1
azure-common==1.1.28
azure-core==1.26.3
azure-functions==1.13.2
azure-storage-blob==12.14.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
idna==3.4
isodate==0.6.1
msrest==0.7.1
numpy==1.24.2
oauthlib==3.2.2
pandas==1.5.3
praw==7.6.1
prawcore==2.3.0
pycparser==2.21
pyodbc==4.0.35
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.5.0
update-checker==0.18.0
urllib3==1.26.14
websocket-client==1.5.1
